requirements.txt
Django==3.0.6
django-cors-headers==3.3.0
django-filter==2.2.0
djangorestframework==3.11.0
gunicorn==20.0.4
psycopg2-binary==2.8.5
